vertex coordinate changed to long

+++ src/com/teamdev/Vertex.java
@@ -6,14 +6,14 @@ package com.teamdev;
 public class Vertex {
 
     // x
-    public int nX; // nominator
-    public int dX; // denominator
+    public long nX; // nominator
+    public long dX; // denominator
 
     // y
-    public int nY; // nominator
-    public int dY; // denominator
+    public long nY; // nominator
+    public long dY; // denominator
 
-    public Vertex(int nX, int dX, int nY, int dY) {
+    public Vertex(long nX, long dX, long nY, long dY) {
         this.nX = nX;
         this.dX = dX;
         this.nY = nY;
@@ -29,30 +29,30 @@ public class Vertex {
 
         i = xLine.indexOf('/');
         if( i == -1 ) {
-            nX = Integer.parseInt(xLine);
+            nX = Long.parseLong(xLine);
             dX = 1;
         }
         else {
-            nX = Integer.parseInt(xLine.substring(0, i).trim());
-            dX = Integer.parseInt(xLine.substring(i+1).trim());
+            nX = Long.parseLong(xLine.substring(0, i).trim());
+            dX = Long.parseLong(xLine.substring(i+1).trim());
         }
 
         i = yLine.indexOf('/');
         if( i == -1 ) {
-            nY = Integer.parseInt(yLine);
+            nY = Long.parseLong(yLine);
             dY = 1;
         }
         else {
-            nY = Integer.parseInt(yLine.substring(0, i).trim());
-            dY = Integer.parseInt(yLine.substring(i+1).trim());
+            nY = Long.parseLong(yLine.substring(0, i).trim());
+            dY = Long.parseLong(yLine.substring(i+1).trim());
         }
     }
 
     public float getFloatX() {
-        return (float)nX / (float)dX;
+        return (float) ((double)nX / (double) dX);
     }
 
     public float getFloatY() {
-        return (float)nY / (float)dY;
+        return (float) ((double)nY / (double) dY);
     }
 }
